feat: parametrized offline message widget
